Fix output redirection of 'grub --batch' again
grub can take as an argument the device on which it will operate.  That was
the initial behaviour: /dev/null was the device to be operated, as we are
only interested in the GRUB shell output.  So '>' was not missing.

In order to prevent further misreading, we now use '--device-map=/dev/null'
like grub-md5-crypt does.

r49789