Simplify grub package selection code.
This code had developed into a real confusing mess, where grub_package
was set to grub and then this was (almost) always overridden to grub-pc.

As best I can tell, this new code should have the same effect
as the old: grub2 is used by default except for the two special cases
of sataraid and multipath. (Based on #442382 being closed,
grub2 could probably be used for multipath as well now.)

The special cases for grub-yeeloong grub-ieee1275 remain;
as does the warning for those about grub not being mature
on those platforms.