Remove code that relies on target fstab containing /proc, which it no longer does.
The /proc mounting in here is just a horrible mess.

First there's code that runs in-target. Which handles mounting, and
umounting proc on its own. (It nicely only umounts it if it was not mounted
before.)

Then there was code that assumed mount /proc would work in target .. which
it did until recently when partman-target stopped putting /proc in the
fstab as it's otherwise mounted these days.

And then there's this new ensure_proc_in_target, which will tend be be
defeated, sometimes, by the manual umount /proc called elsewhere in
grub-installer. (And which also, it should be noted, leaves /proc mounted
on exit.)

I have not entirely fixed this horror, but I did make it consistently
use ensure_proc_in_target for its manual mounting, and I added one
umount of /proc on (successful) exit.

A proper fix would seem to need a in-target --no-eat-stdin that was used
consistently where all the chroots are in grub-installer.