Fix SG_IO ioctl to a partition
Do not send SG_IO ioctl to the underline device if it is a partion.

It is even better to remap the SG_IO to the whole device. However it is
not easy (find the device path of the whole device from a fd opened for
a partition).

Before:
[   85.563402] zfs-fuse: sending ioctl 2285 to a partition!
[   85.578762] zfs-fuse: sending ioctl 2285 to a partition!
[   85.578771] zfs-fuse: sending ioctl 2285 to a partition!
[  115.702088] scsi_verify_blk_ioctl: 14 callbacks suppressed
[  115.702097] zfs-fuse: sending ioctl 2285 to a partition!
[  115.702105] zfs-fuse: sending ioctl 2285 to a partition!

After:
no warning