Add armmp subarch
The other kernel flavours have been removed in the 3.11 kernel, even in 3.10
(the kernel currently used by di) the armmp flavour supports mx5 and vexpress.

This new subarch incorporates all of the mediums supported by the existing mx5
and vexpress subarches.

When di switches to 3.11 or newer the mx5 and vexpress subarches will need to
be removed.