arm: include dtb files for netboot.
More and more platforms require a device tree to boot. We include them in the
kernel package and the udeb, export them into the installer output as well.

This introduces a new "medium" to the relevant architectures. This isn't a true
"medium", but since DTBs are common to the netboot/ network-console etc this
puts them at the top level.

This is enabled for armhf and armel/kirkwood. The other armel subarchs don't
appear to produce any device tree.