Keep Linux modules.builtin file in the initrd
modules.builtin is generated at kernel build time and can't be
regenerated from the installed kernel or modules.  If it's missing
when depmod runs again within the installer, depmod will generate an
empty modules.builtin.bin and issue a warning.  modprobe will also
fail noisily if asked to load a module that is actually built-in.