Provide a PXE bootable grub.efi in the netboot flavour for amd64 and arm64.
This provides a suitable grub binary configured to look at PXE for
configuration, plus makes the usual modules etc and a grub.cfg available in the
netboot directory.

arm64 gains the same netboot.tar.gz as x86 already has, as well as the expected
unpacked tree in the output.

This touches x86.cfg but i386.cfg sets GRUB_EFI=n so there is no change there,
only for amd64.

Tested (slowly!) with qemu-system-x86_64 and qemu-system-arm64.